requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-image>=0.20.0
scipy>=1.10.0
Pillow>=9.0.0